auth-service/requirements.txt
Flask==2.3.2
PyJWT==2.6.0

